+++ src/buffer.cpp
@@ -132,3 +132,27 @@ SYZ_CAPI syz_ErrorCode syz_createBufferFromStream(syz_Handle *out, syz_Handle co
 	return 0;
 	SYZ_EPILOGUE
 }
+
+SYZ_CAPI syz_ErrorCode syz_bufferGetChannels(unsigned int *out, syz_Handle buffer) {
+	SYZ_PROLOGUE
+	auto b = fromC<Buffer>(buffer);
+	*out = b->getChannels();
+	return 0;
+	SYZ_EPILOGUE
+}
+
+SYZ_CAPI syz_ErrorCode syz_bufferGetLengthInSamples(unsigned int *out, syz_Handle buffer) {
+	SYZ_PROLOGUE
+	auto b = fromC<Buffer>(buffer);
+	*out = b->getLength();
+	return 0;
+	SYZ_EPILOGUE
+}
+
+SYZ_CAPI syz_ErrorCode syz_bufferGetLengthInSeconds(double *out, syz_Handle buffer) {
+	SYZ_PROLOGUE
+	auto b = fromC<Buffer>(buffer);
+	*out = b->getLength() / (double) config::SR;
+	return 0;
+	SYZ_EPILOGUE
+}